:recycle: Refactor: User Api 위치 이동
<body>
- :external-api:/api/user
  - 기존의 형태에서 domain에 대한 역할 위임이 이루어졌기에 더 보기 편하게 변경

<footer>
- 관련: #139
