+++ application/external-api/src/main/java/com/namo/spring/application/external/api/user/converter/UserResponseConverter.java
@@ -1,8 +1,8 @@
-package com.namo.spring.application.external.domain.user.application.converter;
+package com.namo.spring.application.external.api.user.converter;
 
 import java.util.List;
 
-import com.namo.spring.application.external.domain.user.ui.dto.UserResponse;
+import com.namo.spring.application.external.api.user.dto.UserResponse;
 
 public class UserResponseConverter {
 	private UserResponseConverter() {
