+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/application/MoimFacade.java
@@ -19,7 +19,7 @@ import com.namo.spring.application.external.domain.group.domain.Moim;
 import com.namo.spring.application.external.domain.group.domain.MoimAndUser;
 import com.namo.spring.application.external.domain.group.ui.dto.GroupRequest;
 import com.namo.spring.application.external.domain.group.ui.dto.GroupResponse;
-import com.namo.spring.application.external.domain.user.application.impl.UserService;
+import com.namo.spring.application.external.api.user.service.UserService;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.application.external.global.common.constant.FilePath;
 import com.namo.spring.application.external.global.utils.FileUtils;
