+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/application/MoimMemoFacade.java
@@ -24,7 +24,7 @@ import com.namo.spring.application.external.domain.group.domain.MoimScheduleAndU
 import com.namo.spring.application.external.domain.group.ui.dto.GroupDiaryRequest;
 import com.namo.spring.application.external.domain.group.ui.dto.GroupDiaryResponse;
 import com.namo.spring.application.external.domain.group.ui.dto.GroupScheduleRequest;
-import com.namo.spring.application.external.domain.user.application.impl.UserService;
+import com.namo.spring.application.external.api.user.service.UserService;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.application.external.global.common.constant.FilePath;
 import com.namo.spring.application.external.global.utils.FileUtils;
