+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/application/CategoryFacade.java
@@ -14,7 +14,7 @@ import com.namo.spring.application.external.domain.individual.domain.Palette;
 import com.namo.spring.application.external.domain.individual.ui.dto.CategoryRequest;
 import com.namo.spring.application.external.domain.individual.ui.dto.CategoryResponse;
 
-import com.namo.spring.application.external.domain.user.application.impl.UserService;
+import com.namo.spring.application.external.api.user.service.UserService;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 
 import lombok.RequiredArgsConstructor;
